+++ src/components/SqlTable/index.ts
@@ -1,1 +1,1 @@
-export { default } from "./SqlTable";
+export { default, type FilterCache, type RememberedFilters } from "./SqlTable";